client: store control groups on cl like selection
Move numbered groups off statics in cl_input_w3.c onto cl.groups.
The group command is registered with the rest of client input and
clears on map load / CL_ClearState.
