+++ client/cl_control_groups.h
@@ -9,4 +9,6 @@
 DWORD CL_ControlGroupAppendUnique(DWORD *group, DWORD count, DWORD capacity,
                                   DWORD const *ids, DWORD num_ids);
 
+void CL_ControlGroupsReset(void);
+
 #endif
